// src/models/workout.ts
import { Exercise } from "./exercise";

/**
 * This is a specific instantiation of a workout, generated by the user. It corresponds to the
 * user-workout table and the user-exercises associated with it.
 *
 * This is used whenever information about past workouts are needed, and when a new workout
 * is being logged.
 */
  
import { WorkoutType } from "./workout-type";
export class Workout {
  public order: number;
  public date: string;
  public type: WorkoutType;
  public exercises: Exercise[];

  constructor(
    type: WorkoutType,
    order: number,
    exercises: Exercise[],
    date: string
  ) {
    this.type = type;
    this.exercises = exercises;
    this.order = order;
    this.date = date;
  }
}